// frontend/libs/canvasSpreadsheet/src/lib/types/types.ts
export type Edges = {
  startRow: number;
  endRow: number;
  startCol: number;
  endCol: number;
};

export type ViewportEdges = Edges;
export type SelectionEdges = Edges;
// ...
export type GridTable = {
  startRow: number;
  startCol: number;
  endCol: number;
  endRow: number;
  tableName: string;
  chartType?: ChartType;
  isTableNameHeaderHidden: boolean;
  isTableFieldsHeaderHidden: boolean;
  isTableHorizontal: boolean;
  totalSize: number;
  hasKeys: boolean;
  isManual: boolean;
  highlightType?: 'DIMMED' | 'NORMAL' | 'HIGHLIGHTED' | undefined;
  note?: string;
  fieldNames: string[];
};

export type GridField = {
  fieldName: string;
  expression: string;
  isKey: boolean;
  isDim: boolean;
  isNested: boolean;
  isPeriodSeries: boolean;
  isDynamic: boolean;
  isFiltered: boolean;
  conditionFilters?: FieldConditionFiltersResult;
  filterExpression?: string;
  filterControlRef?: {
    controlTableName: string;
    controlFieldName: string;
  } | null;
  totalFieldTypes?: TotalType[];
  sort: FieldSortOrder;
  isFieldUsedInSort: boolean;
  type: ColumnDataType;
  format: ColumnFormat | undefined;
  referenceTableName?: string;
  note?: string;
  hasError: boolean;
  errorMessage?: string;
  highlightType?: 'DIMMED' | 'NORMAL' | 'HIGHLIGHTED' | undefined;
  isInput: boolean;
  isImport: boolean;
  isIndex: boolean;
  isControl: boolean;
  controlType?: ControlType;
  isDescription: boolean;
  descriptionField?: string;
  dataLength: number;
  hasOverrides: boolean;
  isAIFunctions: boolean;
  isLoading?: boolean;
};

export type GridCell = {
  table?: GridTable;
  field?: GridField;
  value?: string;
  displayValue?: string;

  hasError?: boolean;
  errorMessage?: string;

  totalIndex?: number;
  totalType?: TotalType;
  totalExpression?: string;

  row: number;
  col: number;
  dataIndex?: number;

  overrideIndex?: number;
  overrideValue?: OverrideValue;
  isOverride?: boolean;
  overrideAIFunctions?: boolean;

  isUrl?: boolean;

  isTableHeader?: boolean;
  isFieldHeader?: boolean;

  isRightAligned?: boolean;

  startCol: number;
  endCol: number;

  startGroupColOrRow: number;
  endGroupColOrRow: number;
};

export type RowData = { [col: string]: GridCell };

export type GridData = {
  [row: string]: RowData;
};